move blacklist options to movies and mics only
