+++ pyworkflow/em/protocol/protocol_import/base.py
@@ -28,12 +28,14 @@ import os
 from os.path import join
 from glob import glob
 import re
-from datetime import timedelta, datetime
+from datetime import datetime
 
 import pyworkflow.utils as pwutils
 import pyworkflow.protocol.params as params
 from pyworkflow.utils.path import expandPattern, copyFile, createAbsLink
 from pyworkflow.em.protocol import EMProtocol
+
+
 
 class ProtImport(EMProtocol):
     """ Base class for other all Import protocols. """
@@ -50,25 +52,24 @@ class ProtImportFiles(ProtImport):
         (_importFile(fileName, fileId))
     """
     IMPORT_FROM_FILES = 0
-    BLACKLIST_REGEXPS = 0
 
-    # --------------------------- DEFINE param functions ------------------------
+    #--------------------------- DEFINE param functions ------------------------
     def _defineParams(self, form):
         importChoices = self._getImportChoices()
         filesCondition = self._getFilesCondition()
 
         form.addSection(label='Import')
 
-        if len(importChoices) > 1:  # not only from files
+        if len(importChoices) > 1: # not only from files
             form.addParam('importFrom', params.EnumParam,
                           choices=importChoices, default=self._getDefaultChoice(),
                           label='Import from',
                           help='Select the type of import.')
         else:
             form.addHidden('importFrom', params.EnumParam,
-                           choices=importChoices, default=self.IMPORT_FROM_FILES,
-                           label='Import from',
-                           help='Select the type of import.')
+                          choices=importChoices, default=self.IMPORT_FROM_FILES,
+                          label='Import from',
+                          help='Select the type of import.')
         form.addParam('filesPath', params.PathParam,
                       condition=filesCondition,
                       label="Files directory",
@@ -112,33 +113,33 @@ class ProtImportFiles(ProtImport):
         form.addSection('Streaming')
 
         form.addParam('dataStreaming', params.BooleanParam, default=False,
-                      label="Process data in streaming?",
-                      help="Select this option if you want import data as it is "
-                           "generated and process on the fly by next protocols. "
-                           "In this case the protocol will keep running to check "
-                           "new files and will update the output Set, which can "
-                           "be used right away by next steps.")
+              label="Process data in streaming?",
+              help="Select this option if you want import data as it is "
+                   "generated and process on the fly by next protocols. "
+                   "In this case the protocol will keep running to check "
+                   "new files and will update the output Set, which can "
+                   "be used right away by next steps.")
 
         form.addParam('timeout', params.IntParam, default=43200,
-                      condition='dataStreaming',
-                      label="Timeout (secs)",
-                      help="Interval of time (in seconds) after which, if no new file "
-                           "is detected, the protocol will end. When finished, "
-                           "the output Set will be closed and no more data will be "
-                           "added to it. \n"
-                           "Note 1:  The default value is  high (12 hours) to avoid "
-                           "the protocol finishes during the aqcuisition of the "
-                           "microscpe. You can also stop it from right click and press "
-                           "STOP_STREAMING.\n"
-                           "Note 2: If you're using individual frames when importing "
-                           "movies, the timeout won't be refreshed until a whole "
-                           "movie is stacked.")
+              condition='dataStreaming',
+              label="Timeout (secs)",
+              help="Interval of time (in seconds) after which, if no new file "
+                   "is detected, the protocol will end. When finished, "
+                   "the output Set will be closed and no more data will be "
+                   "added to it. \n"
+                    "Note 1:  The default value is  high (12 hours) to avoid "
+                   "the protocol finishes during the aqcuisition of the "
+                   "microscope. You can also stop it from right click and press "
+                   "STOP_STREAMING.\n"
+                   "Note 2: If you're using individual frames when importing "
+                   "movies, the timeout won't be refreshed until a whole "
+                   "movie is stacked.")
 
         form.addParam('fileTimeout', params.IntParam, default=30,
-                      condition='dataStreaming',
-                      label="File timeout (secs)",
-                      help="Interval of time (in seconds) after which, if a file has "
-                           "not changed, we consider it as a new file. \n")
+              condition='dataStreaming',
+              label="File timeout (secs)",
+              help="Interval of time (in seconds) after which, if a file has "
+                   "not changed, we consider it as a new file. \n")
 
         self._defineBlacklistParams(form)
 
@@ -148,60 +149,20 @@ class ProtImportFiles(ProtImport):
         """
         pass
 
-    def _getBlacklistSetClass(self):
-        """ Returns the class to be blacklisted by this protocol.
-        """
-        return "SetOfImages"
-
-    def _defineBlacklistParams(self, form):
-        """ Options to blacklist certain items when launching the
-        import protocol.
-        """
-        form.addSection(label="Blacklist")
-        form.addParam("blacklistSet", params.PointerParam,
-                      pointerClass=self._getBlacklistSetClass(),
-                      allowsNull=True,
-                      label="Blacklist Set",
-                      help="Files on this set will not be imported")
-        form.addParam('blacklistDateFrom', params.StringParam,
-                      label="Blacklist from date",
-                      allowsNull=True,
-                      help="Files acquired after this date will not be imported. "
-                           "Must follow format: YYYY-mm-dd HH:MM:SS \n"
-                           "e.g: 2019-01-14 14:18:05")
-        form.addParam('blacklistDateTo', params.StringParam,
-                      label="Blacklist to date",
-                      allowsNull=True,
-                      help="Files acquired before this date will not be imported. "
-                           "Must follow format: YYYY-mm-dd HH:MM:SS \n"
-                           "e.g: 2019-01-14 14:18:05")
-        form.addParam('useRegexps', params.EnumParam,
-                      default=self.BLACKLIST_REGEXPS,
-                      choices=['RegExps', 'File names'],
-                      label='Blacklist file type',
-                      help="Choose RegExp if the black list file contains regular expressions. Set to File Names if "
-                           "the black list file contains file names.")
-        form.addParam('blacklistFile', params.FileParam,
-                      label="Blacklist File",
-                      allowsNull=True,
-                      help="Blacklist everything included in this file. If Use RegExps is True,"
-                           "lines will be interpreted as regular expressions. E.g: \n"
-                           "(.*)GRID_0[1-5](.*)\n"
-                           "(.*)/GRID_10/Falcon_2019_01_14-16_(.*)\n"
-                           "If Use RegExps is False, lines will be interpreted as file names. E.g.\n"
-                           "/path/to/GRID_10/Falcon_2019_01_14-16_51_20_0_movie.mrcs\n"
-                           "/path/to/GRID_10/Falcon_2019_01_14-16_55_40_0_movie.mrcs"
-                      )
-
     def _defineAcquisitionParams(self, form):
         """ Override to add options related to acquisition info.
         """
         pass
 
-    def _getDefaultChoice(self):
-        return self.IMPORT_FROM_FILES
+    def _defineBlacklistParams(self, form):
+        """ Override to add options related to blacklist info.
+        """
+        pass
 
-    # --------------------------- INFO functions --------------------------------
+    def _getDefaultChoice(self):
+        return  self.IMPORT_FROM_FILES
+
+    #--------------------------- INFO functions --------------------------------
     def _validate(self):
         errors = []
         if self.importFrom == self.IMPORT_FROM_FILES:
@@ -214,20 +175,9 @@ class ProtImportFiles(ProtImport):
                     errors.append("There are no files matching the pattern %s"
                                   % self.getPattern())
 
-        dates = [self.blacklistDateFrom.get(), self.blacklistDateTo.get()]
-        parsedDates = []
-        for d in dates:
-            if d:
-                try:
-                    parsedDates.append(datetime.strptime(self.blacklistDateTo.get(), "%Y-%m-%d %H:%M:%S"))
-                except ValueError as e:
-                    errors.append("Bad date formatting in blacklist date %s: %s" % (d, e))
-
-        if len(parsedDates) == 2 and parsedDates[0] > parsedDates[1]:
-            errors.append("Wrong blacklist dates: date from must be earlier than date to")
         return errors
 
-    # --------------------------- BASE methods to be overriden ------------------
+    #--------------------------- BASE methods to be overriden ------------------
     def _getImportChoices(self):
         """ Return a list of possible choices
         from which the import can be done.
@@ -242,7 +192,7 @@ class ProtImportFiles(ProtImport):
         """
         return '(importFrom == %d)' % self.IMPORT_FROM_FILES
 
-    # --------------------------- UTILS functions -------------------------------
+    #--------------------------- UTILS functions -------------------------------
     def getPattern(self):
         """ Expand the pattern using environ vars or username
         and also replacing special character # by digit matching.
@@ -263,11 +213,11 @@ class ProtImportFiles(ProtImport):
             g = match.group(1)
             n = len(g)
             # prepare regex pattern - place ids, handle *, handle ?
-            idregex = pattern.replace(g, '(%s)' % ('[0-9]' * n))
-            idregex = idregex.replace('*', '.*')
+            idregex = pattern.replace(g, '(%s)' % ('[0-9]'*n))
+            idregex = idregex.replace('*','.*')
             idregex = idregex.replace('?', '.')
             self._idRegex = re.compile(idregex)
-            pattern = pattern.replace(g, '[0-9]' * n)
+            pattern = pattern.replace(g, '[0-9]'*n)
 
         return pattern
 
@@ -306,86 +256,9 @@ class ProtImportFiles(ProtImport):
 
         return delta < fileTimeout
 
-    def _getUniqueFileName(self, fileName):
-        """To be overwritten by subclasses"""
-        return os.path.basename(fileName)
-
-    def getItemsToBlacklistFromFile(self):
-        if not hasattr(self, '_fileItemsToBlacklist'):
-            blacklistfile = self.blacklistFile.get()
-            blacklistItems = set()
-            if blacklistfile:
-                with open(blacklistfile, 'r') as f:
-                    for blacklistedItem in f:
-                        blacklistedItem = blacklistedItem.strip()
-                        blacklistItems.add(blacklistedItem)
-            self._fileItemsToBlacklist = blacklistItems
-
-        return self._fileItemsToBlacklist
-
-    def getBlacklistedItems(self):
-        if not hasattr(self, '_blacklistedItems'):
-            self._blacklistedItems = set()
-        return self._blacklistedItems
-
     def isBlacklisted(self, fileName):
-        # check if already blacklisted
-        blacklistedItems = self.getBlacklistedItems()
-        if fileName in blacklistedItems:
-            return True
-
-        # Blacklisted by set
-        blacklistSet = self.blacklistSet.get()
-        if blacklistSet is not None:
-            for img in blacklistSet:
-                blacklistFileName = img.getFileName()
-                if ((os.path.islink(blacklistFileName)
-                     and fileName == os.readlink(blacklistFileName))
-                        or (self._getUniqueFileName(fileName) == os.path.basename(blacklistFileName))):
-                    print("Blacklist warning: %s is blacklisted by the input set" % fileName)
-                    blacklistedItems.add(fileName)
-                    return True
-
-        # Blacklisted by date
-        blacklistDateFrom = self.blacklistDateFrom.get()
-        blacklistDateTo = self.blacklistDateTo.get()
-        doDateBlacklist = blacklistDateFrom is not None and blacklistDateTo is not None
-        if doDateBlacklist:
-            fileDate = datetime.fromtimestamp(os.path.getmtime(fileName))
-            if blacklistDateFrom:
-                parsedDateFrom = datetime.strptime(blacklistDateFrom, "%Y-%m-%d %H:%M:%S")
-                if blacklistDateTo:
-                    parsedDateTo = datetime.strptime(blacklistDateTo, "%Y-%m-%d %H:%M:%S")
-                    if parsedDateFrom <= fileDate <= parsedDateTo:
-                        print("Blacklist warning: %s is blacklisted by date" % fileName)
-                        blacklistedItems.add(fileName)
-                        return True
-                else:
-                    if parsedDateFrom <= fileDate:
-                        print("Blacklist warning: %s is blacklisted by date" % fileName)
-                        blacklistedItems.add(fileName)
-                        return True
-
-            elif blacklistDateTo:
-                parsedDateTo = datetime.strptime(blacklistDateTo, "%Y-%m-%d %H:%M:%S")
-                if fileDate <= parsedDateTo:
-                    print("Blacklist warning: %s is blacklisted by date" % fileName)
-                    blacklistedItems.add(fileName)
-                    return True
-
-        # Blacklisted by file
-        items2blacklist = self.getItemsToBlacklistFromFile()
-        for item2blacklist in items2blacklist:
-            if self.useRegexps.get() == self.BLACKLIST_REGEXPS:
-                if re.match(item2blacklist, fileName):
-                    print("Blacklist warning: %s matched blacklist regexp %s"
-                          % (fileName, item2blacklist))
-                    blacklistedItems.add(fileName)
-                    return True
-            elif fileName in item2blacklist:
-                print("Blacklist warning: %s is blacklisted " % fileName)
-                blacklistedItems.add(fileName)
-                return True
+        """ Overwrite in subclasses """
+        return False
 
     def iterFiles(self):
         """ Iterate through the files matched with the pattern.
@@ -410,3 +283,5 @@ class ProtImportFiles(ProtImport):
                 fileId = None
 
             yield fileName, fileId
+
+
